requirements.txt
pandas==1.0.5
tornado==6.0.4
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==20.0.4
python-dotenv==0.14.0
mysqlclient==2.0.1
fastparquet==0.4.1
drf-yasg==1.17.1
